v1.0
The code coverage is the barely minimum to allow the forward RUS problem to be solved.

Index variables are incremented, because Julia counts from 1, not from 0.

The crystal identifiers are not integer-based as in PALab/RUS (ns, hextype),
but type-based, and the julia dispatch system takes care of it;
the calls precompile, so this should not result in a performance penalty.

The shape (spherical, cylindrical, rectangular) is still index based, but goes from 1 instead of from 0.